insightx_ingestion/requirements.txt
fastapi
uvicorn[standard]
supabase
pgvector
groq
httpx
python-dotenv
pydantic>=2.0
gTTS
newspaper3k
pydantic-settings
yfinance
lxml[html-clean]
